+++ src/components/EmptyGarage.jsx
@@ -4,7 +4,7 @@
 
 // import { useState } from "react";
 import { NavLink } from "react-router-dom";
-import "../../App.css";
+import "../App.css";
 
 export default function emptyGarage() {
   // const [isEmpty, setIsEmpty] = useState("");
